# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "API Adèle Délice",
    "description": "API REST pour le restaurant Adèle Délice : gestion des menus, plats, commandes et réservations"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Serveur de développement"
    }
  ],
  "tags": [
    {
      "name": "Catégories",
      "description": "Gestion des catégories de plats"
    },
    {
      "name": "Plats",
      "description": "Gestion des plats"
    },
    {
      "name": "Menus",
      "description": "Gestion des menus"
    },
    {
      "name": "Commandes",
      "description": "Gestion des commandes"
    },
    {
      "name": "Réservations",
      "description": "Gestion des réservations"
    }
  ],
  "paths": {
    "/api/categories": {
      "get": {
        "tags": ["Catégories"],
        "summary": "Récupère toutes les catégories actives",
        "responses": {
          "200": {
            "description": "Liste des catégories",
            "content": {
              "application/json": {}
            }
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      },
      "post": {
        "tags": ["Catégories"],
        "summary": "Crée une nouvelle catégorie",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {}
          }
        },
        "responses": {
          "201": {
            "description": "Catégorie créée"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
    "/api/categories/{id}": {
      "get": {
        "tags": ["Catégories"],
        "summary": "Récupère une catégorie par son ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Catégorie trouvée"
          },
          "404": {
            "description": "Catégorie introuvable"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      },
      "put": {
        "tags": ["Catégories"],
        "summary": "Met à jour une catégorie",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {}
          }
        },
        "responses": {
          "200": {
            "description": "Catégorie mise à jour"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      },
      "delete": {
        "tags": ["Catégories"],
        "summary": "Supprime une catégorie",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Catégorie supprimée"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
    "/api/dishes": {
      "get": {
        "tags": ["Plats"],
        "summary": "Récupère tous les plats",
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isPromoted",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isNew",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des plats"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      },
      "post": {
        "tags": ["Plats"],
        "summary": "Crée un nouveau plat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {}
          }
        },
        "responses": {
          "201": {
            "description": "Plat créé"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
    "/api/dishes/{id}": {
      "get": {
        "tags": ["Plats"],
        "summary": "Récupère un plat par son ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plat trouvé"
          },
          "404": {
            "description": "Plat introuvable"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      },
      "put": {
        "tags": ["Plats"],
        "summary": "Met à jour un plat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {}
          }
        },
        "responses": {
          "200": {
            "description": "Plat mis à jour"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      },
      "delete": {
        "tags": ["Plats"],
        "summary": "Supprime un plat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plat supprimé"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
    "/api/menus": {
      "get": {
        "tags": ["Menus"],
        "summary": "Récupère tous les menus actifs",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["DAILY", "WEEKLY", "SPECIAL"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des menus"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
    "/api/orders": {
      "get": {
        "tags": ["Commandes"],
        "summary": "Récupère toutes les commandes ou une commande par son numéro",
        "parameters": [
          {
            "name": "orderNumber",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des commandes ou une commande"
          },
          "404": {
            "description": "Commande introuvable"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      },
      "post": {
        "tags": ["Commandes"],
        "summary": "Crée une nouvelle commande",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {}
          }
        },
        "responses": {
          "201": {
            "description": "Commande créée"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": ["Commandes"],
        "summary": "Récupère une commande par son ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Commande trouvée"
          },
          "404": {
            "description": "Commande introuvable"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      },
      "put": {
        "tags": ["Commandes"],
        "summary": "Met à jour une commande (changement de statut)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {}
          }
        },
        "responses": {
          "200": {
            "description": "Commande mise à jour"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
    "/api/reservations": {
      "get": {
        "tags": ["Réservations"],
        "summary": "Récupère toutes les réservations",
        "responses": {
          "200": {
            "description": "Liste des réservations"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      },
      "post": {
        "tags": ["Réservations"],
        "summary": "Crée une nouvelle réservation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {}
          }
        },
        "responses": {
          "201": {
            "description": "Réservation créée"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
    "/api/reservations/{id}": {
      "put": {
        "tags": ["Réservations"],
        "summary": "Met à jour une réservation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {}
          }
        },
        "responses": {
          "200": {
            "description": "Réservation mise à jour"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    }
  }
}